refactor: extract human layer factory
